ODE-314 moving refactor tests
