+++ jpo-ode-svcs/src/test/java/us/dot/its/jpo/ode/udp/controller/UdpServiceThreadFactoryTest.java
@@ -1,4 +1,4 @@
-package us.dot.its.jpo.ode.udp;
+package us.dot.its.jpo.ode.udp.controller;
 
 import static org.junit.Assert.assertEquals;
