+++ jpo-ode-svcs/src/test/java/us/dot/its/jpo/ode/udp/trust/TrustSessionTest.java
@@ -1,4 +1,4 @@
-package us.dot.its.jpo.ode.udp;
+package us.dot.its.jpo.ode.udp.trust;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
